--- SplitTests/Integration/streaming/NewStreamingInitTest.swift
@@ -1,80 +1,0 @@
-//
-// StreamingInitTest.swift
-// Split
-//
-// Created by Javier L. Avrudsky on 14-Oct-2020.
-// Copyright (c) 2020 Split. All rights reserved.
-//
-
-import XCTest
-@testable import Split
-
-class StreamingInitTest: XCTestCase {
-    var httpClient: HttpClient!
-    let apiKey = IntegrationHelper.dummyApiKey
-    let userKey = IntegrationHelper.dummyUserKey
-    var isSseAuthHit = false
-    var isSseHit = false
-
-    override func setUp() {
-        let session = HttpSessionMock()
-        let reqManager = HttpRequestManagerTestDispatcher(dispatcher: buildTestDispatcher())
-        httpClient = DefaultHttpClient(session: session, requestManager: reqManager)
-    }
-
-    func testInit() {
-
-        let splitConfig: SplitClientConfig = SplitClientConfig()
-        splitConfig.featuresRefreshRate = 30
-        splitConfig.segmentsRefreshRate = 30
-        splitConfig.impressionRefreshRate = 30
-        splitConfig.sdkReadyTimeOut = 60000
-        splitConfig.eventsPerPush = 10
-        splitConfig.eventsQueueSize = 100
-        splitConfig.eventsPushRate = 5
-
-        let key: Key = Key(matchingKey: userKey)
-        let factory = DefaultSplitFactoryBuilder().setHttpClient(httpClient).setApiKey(apiKey).setKey(key)
-            .setConfig(splitConfig).build()!
-
-        let client = factory.client
-
-        let sdkReadyExpectation = XCTestExpectation(description: "SDK READY Expectation")
-        var timeOutFired = false
-        var sdkReadyFired = false
-
-        client.on(event: SplitEvent.sdkReady) {
-            sdkReadyFired = true
-            sdkReadyExpectation.fulfill()
-        }
-
-        client.on(event: SplitEvent.sdkReadyTimedOut) {
-            timeOutFired = true
-            sdkReadyExpectation.fulfill()
-        }
-
-        wait(for: [sdkReadyExpectation], timeout: 5)
-
-        XCTAssertTrue(sdkReadyFired)
-        XCTAssertFalse(timeOutFired)
-        XCTAssertTrue(isSseAuthHit)
-
-    }
-
-    private func buildTestDispatcher() -> HttpClientTestDispatcher {
-        return { request in
-            switch request.url.absoluteString {
-            case let(urlString) where urlString.contains("splitChanges"):
-                return TestDispatcherResponse(code: 200, data: Data(IntegrationHelper.emptySplitChanges.utf8))
-            case let(urlString) where urlString.contains("mysegments"):
-                return TestDispatcherResponse(code: 200, data: Data(IntegrationHelper.emptyMySegments.utf8))
-            case let(urlString) where urlString.contains("auth"):
-                self.isSseAuthHit = true
-                return TestDispatcherResponse(code: 200, data: Data(IntegrationHelper.dummySseResponse().utf8))
-            default:
-                return TestDispatcherResponse(code: 500)
-            }
-        }
-    }
-
-}
