--- SplitTests/Integration/streaming/StreamingConnFail5xx.swift
@@ -1,100 +1,0 @@
-//
-//  StreamingConnFail5xx.swift
-//  SplitTests
-//
-//  Created by Javier L. Avrudsky on 15/10/2020.
-//  Copyright © 2020 Split. All rights reserved.
-//
-
-import XCTest
-@testable import Split
-
-class StreamingConnFail5xxTest: XCTestCase {
-    var httpClient: HttpClient!
-    let apiKey = IntegrationHelper.dummyApiKey
-    let userKey = IntegrationHelper.dummyUserKey
-    var isSseConnected = false
-    var streamingBinding: TestStreamResponseBinding?
-    let sseExp = XCTestExpectation(description: "Sse conn")
-    let kMaxSseConnRetries = 3
-    var sseConnHits = 0
-
-    override func setUp() {
-        let session = HttpSessionMock()
-        let reqManager = HttpRequestManagerTestDispatcher(dispatcher: buildTestDispatcher(),
-                                                          streamingHandler: buildStreamingHandler())
-        httpClient = DefaultHttpClient(session: session, requestManager: reqManager)
-    }
-
-    func testInit() {
-        let splitConfig: SplitClientConfig = SplitClientConfig()
-        splitConfig.featuresRefreshRate = 999999
-        splitConfig.segmentsRefreshRate = 999999
-        splitConfig.impressionRefreshRate = 999999
-        splitConfig.sdkReadyTimeOut = 60000
-        splitConfig.eventsPushRate = 999999
-
-        let key: Key = Key(matchingKey: userKey)
-        let builder = DefaultSplitFactoryBuilder()
-        _ = builder.setHttpClient(httpClient)
-        _ = builder.setReachabilityChecker(ReachabilityMock())
-        let factory = builder.setApiKey(apiKey).setKey(key)
-            .setConfig(splitConfig).build()!
-
-        let client = factory.client
-
-        let sdkReadyExpectation = XCTestExpectation(description: "SDK READY Expectation")
-        var timeOutFired = false
-        var sdkReadyFired = false
-
-        client.on(event: SplitEvent.sdkReady) {
-            sdkReadyFired = true
-            sdkReadyExpectation.fulfill()
-        }
-
-        client.on(event: SplitEvent.sdkReadyTimedOut) {
-            timeOutFired = true
-            sdkReadyExpectation.fulfill()
-        }
-
-        wait(for: [sdkReadyExpectation, sseExp], timeout: 20)
-
-        XCTAssertTrue(sdkReadyFired)
-        XCTAssertFalse(timeOutFired)
-        XCTAssertEqual(3, sseConnHits)
-        XCTAssertTrue(isSseConnected)
-    }
-
-    private func buildTestDispatcher() -> HttpClientTestDispatcher {
-        return { request in
-            switch request.url.absoluteString {
-            case let(urlString) where urlString.contains("splitChanges"):
-                return TestDispatcherResponse(code: 200, data: Data(IntegrationHelper.emptySplitChanges.utf8))
-            case let(urlString) where urlString.contains("mySegments"):
-                return TestDispatcherResponse(code: 200, data: Data(IntegrationHelper.emptyMySegments.utf8))
-            case let(urlString) where urlString.contains("auth"):
-                return TestDispatcherResponse(code: 200, data: Data(IntegrationHelper.dummySseResponse().utf8))
-            default:
-                return TestDispatcherResponse(code: 500)
-            }
-        }
-    }
-
-    private func buildStreamingHandler() -> TestStreamResponseBindingHandler {
-        return { request in
-            self.sseConnHits+=1
-            if self.sseConnHits < self.kMaxSseConnRetries {
-                let bind = TestStreamResponseBinding.createFor(request: request, code: 500)
-                DispatchQueue.global().asyncAfter(deadline: .now() + 0.3) {
-                    bind.complete(error: nil)
-                }
-                return bind
-            }
-            self.streamingBinding = TestStreamResponseBinding.createFor(request: request, code: 200)
-            self.isSseConnected = true
-            self.sseExp.fulfill()
-            return self.streamingBinding!
-        }
-    }
-
-}
